indexにまとめた、menuとshopのバリデーション追加、スクロール追加
