+++ client/components/Menu.tsx
@@ -152,5 +152,5 @@ const Container = styled.div`
 `;
 
 const PageNum = styled.div`
-  margin: 2rem 0;
+  margin-bottom: 2rem;
 `;
